pyproject.toml
[project]
name = "getshitdonebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.68.2",
    "python-telegram-bot==22.0",
]
